added Day Activity Card to Dashboard

+++ app/src/main/java/com/bruhascended/fitapp/ui/activityjournal/ActivityJournalRecyclerAdapter.kt
@@ -106,7 +106,7 @@ class ActivityJournalRecyclerAdapter (
 
             textviewCalories.text = mContext.getString(
                 calorie_count,
-                separatorInfo.totalCalories.toString()
+                separatorInfo.totalCalories.toInt().toString()
             )
             // TODO: Set Using User Preference
             progressbarCalories.apply {
